+++ src/operation/helper/set-global.ts
@@ -1,4 +1,4 @@
-import { getGlobals } from './get-globals';
+import { getGlobals } from './globals';
 
 export function setGlobal(name: string, value: any) {
   const cache = getGlobals();